Refactor storm model
